backend/requirements.txt
fastapi
uvicorn[standard]
pyserial
aiosqlite
pyyaml
